fix compilation warning in debug mode
git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/soma/soma-io/trunk@82693 613412ce-1361-4b7a-93bf-a3c75af1bae4

+++ src/somagis/image/gisimagereader_d.h
@@ -263,7 +263,7 @@ namespace soma {
         if( ( readout = readBlock( target, len ) ) != (long) len ) {
 #ifdef CARTO_DEBUG
           std::cerr << "readBlock( failed at ( ";
-          for (int d=0; d < ndim; ++d){
+          for (size_t d=0; d < ndim; ++d){
               if (d > 0)
                   std::cerr << ",";
               std::cerr << carto::toString(it.position()[d]);
